requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
Django==4.2.11
djangorestframework==3.15.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.4.1
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
libretranslatepy==2.1.1
lxml==5.2.1
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.4
typing-extensions==4.11.0
urllib3==2.2.1
webencodings==0.5.1
